Split networkserver functionality into separate files
