+++ core/networkserver/downlink_test.go
@@ -1,0 +1,86 @@
+// Copyright © 2016 The Things Network
+// Use of this source code is governed by the MIT license that can be found in the LICENSE file.
+
+package networkserver
+
+import (
+	"testing"
+
+	pb_broker "github.com/TheThingsNetwork/ttn/api/broker"
+	"github.com/TheThingsNetwork/ttn/core/networkserver/device"
+	"github.com/TheThingsNetwork/ttn/core/types"
+	"github.com/brocaar/lorawan"
+	. "github.com/smartystreets/assertions"
+)
+
+func TestHandleDownlink(t *testing.T) {
+	a := New(t)
+	ns := &networkServer{
+		devices: device.NewDeviceStore(),
+	}
+
+	appEUI := types.AppEUI(getEUI(1, 2, 3, 4, 5, 6, 7, 8))
+	devEUI := types.DevEUI(getEUI(1, 2, 3, 4, 5, 6, 7, 8))
+	devAddr := getDevAddr(1, 2, 3, 4)
+
+	// Device Not Found
+	message := &pb_broker.DownlinkMessage{
+		AppEui:  &appEUI,
+		DevEui:  &devEUI,
+		Payload: []byte{},
+	}
+	_, err := ns.HandleDownlink(message)
+	a.So(err, ShouldNotBeNil)
+
+	ns.devices.Set(&device.Device{
+		DevAddr: devAddr,
+		AppEUI:  appEUI,
+		DevEUI:  devEUI,
+	})
+
+	// Invalid Payload
+	message = &pb_broker.DownlinkMessage{
+		AppEui:  &appEUI,
+		DevEui:  &devEUI,
+		Payload: []byte{},
+	}
+	_, err = ns.HandleDownlink(message)
+	a.So(err, ShouldNotBeNil)
+
+	fPort := uint8(3)
+	phy := lorawan.PHYPayload{
+		MHDR: lorawan.MHDR{
+			MType: lorawan.UnconfirmedDataDown,
+			Major: lorawan.LoRaWANR1,
+		},
+		MACPayload: &lorawan.MACPayload{
+			FPort: &fPort,
+			FHDR: lorawan.FHDR{
+				FCtrl: lorawan.FCtrl{
+					ACK: true,
+				},
+			},
+		},
+	}
+	bytes, _ := phy.MarshalBinary()
+
+	message = &pb_broker.DownlinkMessage{
+		AppEui:  &appEUI,
+		DevEui:  &devEUI,
+		Payload: bytes,
+	}
+	res, err := ns.HandleDownlink(message)
+	a.So(err, ShouldBeNil)
+
+	var phyPayload lorawan.PHYPayload
+	phyPayload.UnmarshalBinary(res.Payload)
+	macPayload, _ := phyPayload.MACPayload.(*lorawan.MACPayload)
+	a.So(*macPayload.FPort, ShouldEqual, 3)
+	a.So(macPayload.FHDR.DevAddr, ShouldEqual, lorawan.DevAddr{1, 2, 3, 4})
+	a.So(macPayload.FHDR.FCnt, ShouldEqual, 0)                // The first Frame counter is zero
+	a.So(phyPayload.MIC, ShouldNotEqual, [4]byte{0, 0, 0, 0}) // MIC should be set, we'll check it with actual examples in the integration test
+
+	dev, _ := ns.devices.Get(appEUI, devEUI)
+	a.So(dev.FCntDown, ShouldEqual, 1)
+
+}
